Improve DatapathExtensionHost
