+++ hw/chisel/src/main/scala/snax/DataPathExtension/DesignParams.scala
@@ -1,0 +1,10 @@
+package snax.DataPathExtension
+
+class DataPathExtensionParam(
+    val moduleName: String,
+    val userCsrNum: Int,
+    val dataWidth: Int = 512
+) {
+  require(dataWidth > 0)
+  require(userCsrNum >= 0)
+}
